chore: move test on empty multisig hash and add pointer
